added the 'next steps'
after running hardcoreparkour.py, we are left with five mp4 files and one jpg. this new script contains logic for metadata generation and automatic uploading to YT. this is being done separately from the main script because it requires multiple api keys, and some may not know how to authenticate themselves.